Updated source without deprecated castings thanks to old json lib.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/server/gw.h
@@ -53,7 +53,7 @@ namespace K {
       };
       function<void(json*)> helloProduct = [&](json *welcome) {
         *welcome = { {
-          {"exchange", (int)gw->exchange},
+          {"exchange", gw->exchange},
           {"pair", mPair(gw->base, gw->quote)},
           {"minTick", gw->minTick},
           {"environment", ((CF*)config)->argTitle},
@@ -90,8 +90,8 @@ namespace K {
       };
       json serverState() {
         return {
-          {"state",  (int)((QE*)engine)->gwConnectButton},
-          {"status", (int)((QE*)engine)->gwConnectExchange}
+          {"state",  ((QE*)engine)->gwConnectButton},
+          {"status", ((QE*)engine)->gwConnectExchange}
         };
       };
       void handshake(mExchange k) {
